feat : pageNavigator component 생성 및 기존 컴포넌트에서 분리 - 재사용으로 인한 분리
